undo file name change
